Setup config file

+++ main.py
@@ -1,6 +1,9 @@
 import os
+import json
 import shutil
 import getpass
+
+data = json
 
 
 def setup_user(setup: int = 0):
@@ -14,23 +17,42 @@ def setup_user(setup: int = 0):
         except ValueError:
             print('Not valid number')
 
+    user_path = ""
+    path = f'C:\\Users\\{getpass.getuser()}\\Documents'
     if setup == 3:
-        print('Example: \"C:\\Users\\ABC\\Documents\"')
-        path = input('Path:')
-    else:
-        path = f'C:\\Users\\{getpass.getuser()}\\Documents'
+        print('Example: \"C:\\Users\\ABC\\Documents\\E_Learn\"')
+        user_path = input('Path:')
 
-    return path
+    return path, user_path, setup
 
 
 def create_setup_files():
-    path = setup_user()
-    path = os.path.join(path, "E_Learn")
-    try:
-        os.mkdir(path)
-    except FileExistsError:
-        shutil.rmtree(path)
-        os.mkdir(path)
+    path, user_path, setup = setup_user()
+    if setup == 2 or setup == 4:
+        path = os.path.join(path, "E_Learn")
+        try:
+            os.mkdir(path)
+        except FileExistsError:
+            shutil.rmtree(path)
+            os.mkdir(path)
+        init_data = {"config": []}
+        with open(os.path.join(path, "config.json"), 'w') as config:
+            json.dump(init_data, config, indent=4)
+    if setup == 3:
+        config_path = os.path.join(user_path, "config.json")
+        if os.path.isfile(config_path):
+            with open(config_path) as config:
+                global data
+                data = json.load(config)
+                # print(data['config'][1]['address'])
+        path = os.path.join(path, "E_Learn")
+        try:
+            os.mkdir(path)
+        except FileExistsError:
+            shutil.rmtree(path)
+            os.mkdir(path)
+        with open(os.path.join(path, "config.json"), 'w') as config:
+            json.dump(data, config, indent=4)
 
 
 if __name__ == '__main__':
